Create news.py
problem is something to do with input. Maybe input should be reshaped as tensors with the batch size. Having issue though with the sparse tensors. You can access the indices with a .indices and values with .values but can't seem to properly slice the matrix -- it should have dimension n_inputs * 36k (about) because there are 36k words in the dictionary with a tf-idf value.

Tried just putting in numpy arrays, there is an error to do with dimensionality -- indexing with a sequence or something