Add files via upload
Fixed restoring a model to work for all combinations and added a timer to count the time it takes to train a model. 
TODO: Logging with Tensorboard
TODO: Optimise models for best results
Current best: Testing accuracy: 99.2% for a CNN with 2 fully connected layers (Combination 4) 0.001 learning rate, 12 epochs and 128 batch size.